Moved environment factor calculations to seperate submodule
