+++ docs/source/api/soil/env_factors.md
@@ -1,0 +1,23 @@
+---
+jupytext:
+  cell_metadata_filter: -all
+  formats: md:myst
+  main_language: python
+  text_representation:
+    extension: .md
+    format_name: myst
+    format_version: 0.13
+    jupytext_version: 1.13.8
+kernelspec:
+  display_name: vr_python3
+  language: python
+  name: vr_python3
+---
+
+# API documentation for the {mod}`~virtual_rainforest.models.soil.env_factors` module
+
+```{eval-rst}
+.. automodule:: virtual_rainforest.models.soil.env_factors
+    :autosummary:
+    :members:
+```
